Refactoring & test fix
