+++ src/gui/tabs/keys/mod.rs
@@ -1,3 +1,4 @@
 pub mod keys_tab;
 pub mod keygen_subtab;
 pub mod show_xpub_window;
+mod xpub_csv_loader;
