+++ src/gui/tabs/transact/mod.rs
@@ -1,3 +1,9 @@
 pub mod wallet_tab;
 pub mod cold_wallet;
 pub mod hot_wallet;
+pub mod wallet_deprecated;
+pub mod hardware_signing;
+pub mod address_query;
+pub mod prepare_tx;
+pub mod broadcast_tx;
+mod prepared_tx_view;
